Update CLI, fix corruptions dataset dir
